feat: category custom repository로 기능 분리 - #145
- 카테고리 트리를 만드는 기능을
custom category repository로 분리함
- 특정 user의 category를 불러오는 api를
기존 user controller에서 category controller로 옮김
