+++ src/users/users.controller.ts
@@ -22,7 +22,7 @@ import { TransactionInterceptor } from 'src/common/interceptors/transaction.inte
 import { TransactionManager } from 'src/common/transaction.decorator';
 import { EntityManager } from 'typeorm';
 import { EditProfileInput, EditProfileOutput } from './dtos/edit-profile.dto';
-import { LoadPersonalCategoriesOutput } from './dtos/load-personal-categories.dto';
+import { LoadPersonalCategoriesOutput } from '../contents/dtos/load-personal-categories.dto';
 import { LoadPersonalCollectionsOutput } from './dtos/load-personal-collections.dto';
 import {
   LoadFavoritesOutput,
@@ -158,20 +158,20 @@ export class UsersController {
   //   return await this.usersService.loadPersonalCollections(user, +categoryId);
   // }
 
-  @ApiOperation({
-    summary: '자신의 카테고리 목록 조회',
-    description: '자신의 카테고리 목록을 조회하는 메서드',
-  })
-  @ApiOkResponse({
-    description: '카테고리 목록을 반환한다.',
-    type: LoadPersonalCategoriesOutput,
-  })
-  @ApiBearerAuth('Authorization')
-  @UseGuards(JwtAuthGuard)
-  @Get('load-categories')
-  async loadPersonalCategories(
-    @AuthUser() user: User,
-  ): Promise<LoadPersonalCategoriesOutput> {
-    return await this.usersService.loadPersonalCategories(user);
-  }
+  // @ApiOperation({
+  //   summary: '자신의 카테고리 목록 조회',
+  //   description: '자신의 카테고리 목록을 조회하는 메서드',
+  // })
+  // @ApiOkResponse({
+  //   description: '카테고리 목록을 반환한다.',
+  //   type: LoadPersonalCategoriesOutput,
+  // })
+  // @ApiBearerAuth('Authorization')
+  // @UseGuards(JwtAuthGuard)
+  // @Get('load-categories')
+  // async loadPersonalCategories(
+  //   @AuthUser() user: User,
+  // ): Promise<LoadPersonalCategoriesOutput> {
+  //   return await this.usersService.loadPersonalCategories(user);
+  // }
 }
